Add ratelimiter to allow custom hard-coded global ratelimits
This can be usedful for some very large bots that regularly hit the
global ratelimits. It allows them to set a hard-coded limit that fits
their needs and can be shared by multiple shards.

A bot that runs on multiple machines can either use its own
implementation of the Ratelimits that communicates across the network
to ensure that no ratelimits are hit, or just adjust the values and
grant every machine only a part of the available global limit.